+++ src/typedefs.h
@@ -817,6 +817,14 @@ const COMPASUnorderedMap<REMNANT_MASS_PRESCRIPTION, std::string> REMNANT_MASS_PR
     { REMNANT_MASS_PRESCRIPTION::MALTSEV2024,      "MALTSEV2024" }
 };
 
+// maltsev remnant mass prescription variant
+enum class MALTSEV_MODE: int { OPTIMISTIC, BALANCED, PESSIMISTIC };
+const COMPASUnorderedMap<MALTSEV_MODE, std::string> MALTSEV_MODE_LABEL = {
+    { MALTSEV_MODE::OPTIMISTIC,  "OPTIMISTIC" },
+    { MALTSEV_MODE::BALANCED,    "BALANCED"   },
+    { MALTSEV_MODE::PESSIMISTIC, "PESSIMISTIC"},
+};
+
 // response of star to spin-up beyond the Keplerian frequency
 enum class RESPONSE_TO_SPIN_UP: int { TRANSFER_TO_ORBIT, KEPLERIAN_LIMIT, NO_LIMIT };
 const COMPASUnorderedMap<RESPONSE_TO_SPIN_UP, std::string> RESPONSE_TO_SPIN_UP_LABEL = {
@@ -1202,6 +1210,11 @@ typedef struct RLOFProperties {
 
     double       radius1;
     double       radius2;
+
+    double       temperature1;
+    double       temperature2;
+    double       luminosity1;
+    double       luminosity2;
 
     double       starToRocheLobeRadiusRatio1;                                    
     double       starToRocheLobeRadiusRatio2;