+++ src/yaml.h
@@ -349,6 +349,8 @@ namespace yaml {
             "    --fryer-supernova-engine",
             "    --kick-magnitude-distribution",
             "    --kick-direction-distribution",
+            "    --maltsev-fallback",
+            "    --maltsev-mode",
             "    --neutron-star-accretion-in-ce",
             "    --neutron-star-equation-of-state",
             "    --neutrino-mass-loss-BH-formation",